chore: separate string beet into own file
