+++ beet/src/beet.ts
@@ -14,8 +14,10 @@ import {
   numbersTypeMap,
   NumbersTypeMapKey,
 } from './beets/numbers'
+import { StringExports, stringTypeMap, StringTypeMapKey } from './beets/string'
 
 export * from './beets/collections'
+export * from './beets/string'
 export * from './beets/composites'
 export * from './beets/numbers'
 export * from './read-write'
@@ -27,6 +29,7 @@ export * from './types'
  */
 export type BeetTypeMapKey =
   | CollectionsTypeMapKey
+  | StringTypeMapKey
   | CompositesTypeMapKey
   | NumbersTypeMapKey
 
@@ -35,6 +38,7 @@ export type BeetTypeMapKey =
  */
 export type BeetExports =
   | CollectionsExports
+  | StringExports
   | CompositesExports
   | NumbersExports
 
@@ -52,6 +56,7 @@ export const supportedTypeMap: Record<
   }
 > = {
   ...collectionsTypeMap,
+  ...stringTypeMap,
   ...compositesTypeMap,
   ...numbersTypeMap,
 }
